pyproject.toml
[project]
name = "habit-tracking-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Анатолий Сивков",email = "a.sivkov@btc-s.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pytz (>=2025.1,<2026.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.12.12"
flake8-pie = "^0.16.0"

